requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
psycopg2-binary
python-dotenv
email-validator
passlib[argon2]
argon2-cffi
python-jose[cryptography]
python-multipart
fastapi-mail
cloudinary
redis
fastapi-limiter
pytest
pytest-cov
httpx
sphinx
sphinx-rtd-theme
sphinx-autodoc-typehints